requirements.txt
-i https://pypi.org/simple
certifi==2018.11.29
chardet==3.0.4
future==0.17.1
idna==2.8
oauthlib==2.1.0
pymysql==0.9.3
python-twitter==3.5
requests-oauthlib==1.0.0
requests==2.21.0
urllib3==1.24.1
